Add keyable module
